requirements.txt
python-dotenv==1.0.1
langchain-community==0.0.28
langchain==0.1.12
huggingface_hub==0.21.4
pandas==2.2.1
datasets==2.18.0
matplotlib==3.8.3
transformers==4.38.2
sentence-transformers==2.5.1
pymilvus==2.4.0
chromadb==0.3.22